feat(entrar-na-sala) wip término refatoração
